refactor: centralize path utilities in paths.ts
