+++ apps/app/src/lib/ssl.ts
@@ -1,32 +1,11 @@
 import { exec } from "node:child_process";
 import { existsSync, mkdirSync, rmSync } from "node:fs";
-import { join } from "node:path";
 import { promisify } from "node:util";
+import { getSSLDir, getSSLPaths } from "./paths.js";
+
+export { getSSLDir, getSSLPaths } from "./paths.js";
 
 const execAsync = promisify(exec);
-
-function getDataDir(): string {
-  if (process.env.FROST_DATA_DIR) {
-    return process.env.FROST_DATA_DIR;
-  }
-  return join(process.cwd(), "data");
-}
-
-function getSSLBasePath(): string {
-  return process.env.FROST_SSL_PATH || join(getDataDir(), "ssl");
-}
-
-export function getSSLDir(serviceId: string): string {
-  return join(getSSLBasePath(), serviceId);
-}
-
-export function getSSLPaths(serviceId: string): { cert: string; key: string } {
-  const dir = getSSLDir(serviceId);
-  return {
-    cert: join(dir, "server.crt"),
-    key: join(dir, "server.key"),
-  };
-}
 
 export function sslCertsExist(serviceId: string): boolean {
   const { cert, key } = getSSLPaths(serviceId);
